refactor: move testing util types to their own folder
